New flag: Dendarii Mercenaries!
Lois McMaster Bujold is the author of an epic series known to its fans as the Vorkosigan Saga.  The epicenter of the series lies within Miles Vorkosigan, a passionately driven young man who through a series of events forms a military organization which persists throughout the series.

The center of the badge depict the mountain range from which the mercenaries take their name.  The quote is his personal philosophy, which can basically be summarized with the observation that if you just keep running as fast as you can, you will never fall down.